fix(mypy): green the type-check CI job (63→0 errors) — #485 (#488)
* chore(mypy): incremental type-check pass (partial) — issue #485

Partial progress toward greening the `type-check` CI job
(`mypy decision_engine.py common/`), which fails on ~63 pre-existing
errors the mypy config documents as incremental debt.

This commit:
- disables warn_unreachable in the enforced-modules override. It is
  unsound while strict_optional=false — mypy treats Optionals as
  never-None, so legitimate defensive guards (None checks, regex-match
  checks, isinstance checks) are flagged unreachable. Clears 23 such
  false positives without touching the runtime guards. Re-enable with
  strict_optional in the future surgical pass.
- milestone_optionality.py: annotate months/review_months as float and
  evs as List[Dict[str, Any]] (fixes 10 assignment/operator/overload
  errors from dict[str, object] inference).
- rebalance_cost_model.py: estimate_trade_cost_bps returns str bucket
  labels too, so widen the return type to dict[str, Any].
- review_queue.py: annotate groups as dict[str, list].

NOT yet done (tracked, ~30 errors remain across ~18 files): var-annotated
sites, remaining int→float / dict widenings, hash_utils union-attr
type: ignores, and one real latent bug in operator_delivery._is_duplicate
(calls AlertDedupeStore() with no path and non-existent is_duplicate/record
methods — currently swallowed by a bare except, so dedup silently no-ops).
The pytest fixture gap and the lint (repo-wide formatter) job are out of
scope for this commit. type-check will remain red until the remainder lands.


* fix(mypy): resolve remaining type-check errors — issue #485

Completes the incremental type-check pass so `mypy decision_engine.py
common/` is clean (0 errors). All changes are annotation-only or
type-correctness fixes with no runtime behavior change, except two noted
below.

- Widen dict/return types inferred too narrowly (score_to_er, backtest/
  metrics, options_history_massive, options_monitor_v11, competitive_
  intensity, milestone_optionality).
- Add missing annotations (clinical_quality_score, stats/survival,
  stats/cross_sectional, ic_measurement, biotech_hedge_report,
  logging_config, options_diagnostics).
- int→float where a float is assigned (massive_chain_analytics,
  archive_snapshot._human_size).
- decision_engine.compute_sort_contribs return annotation corrected to
  Tuple[Decimal, Dict[str, Decimal]] (already returned Decimal;
  annotation-only — CHANGELOG updated).
- hash_utils / realtime_quote_client: targeted `# type: ignore` matching
  existing patterns. Drop redundant `chain` re-annotation (no-redef) and a
  pre-existing unused import (MAX_SPREAD_PCT) in touched files.

Behavior-adjacent (both align code with its declared contract):
- archive_snapshot "already exists" branch returned str while typed
  -> Path (every other branch returns Path); now returns the Path.
- operator_delivery._is_duplicate called AlertDedupeStore() with no path
  and non-existent methods — raised, swallowed by a bare except, so it
  always returned "not duplicate". Made explicit (return False); real
  dedup wiring is a separate reviewed change (would suppress alerts).

type-check CI job now passes. pytest coverage floor and lint remain
separate (see #485).


---------

+++ decision_engine.py
@@ -2069,7 +2069,7 @@ def compute_sort_contribs(
     alpha_raw: Optional[float] = None,
     catalyst_event_type: str = "",
     catalyst_source: str = "",
-) -> Tuple[float, Dict[str, float]]:
+) -> Tuple[Decimal, Dict[str, Decimal]]:
     """Return sort contribution deltas without recomputing the full sort key.
 
     Returns ``(total_adj, contrib_map)`` where *contrib_map* has an entry
